add--interactive: skip index refresh in reset patch mode
Adds an extra bit to each patch flavour in git-add--interactive,
which determines if the initial index refresh call is required.
It is not required for the reset patch modes, as they do not interact
with the worktree, and do not need up-to-date stat information.

The refresh call can be quite expensive, especially on larger
checkouts or with a freshly created index file.

Signed-off-by: Nika Layzell <nika@thelayzells.com>